Issue #1327: Move test helper code to helpers package.
