+++ app/src/androidTest/java/org/mozilla/tv/firefox/helpers/Assert.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.tv.firefox
+package org.mozilla.tv.firefox.helpers
 
 import android.graphics.Bitmap
 import android.graphics.Color
