+++ app/src/androidTest/java/org/mozilla/tv/firefox/helpers/ext/ViewInteraction.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.tv.firefox.ext
+package org.mozilla.tv.firefox.helpers.ext
 
 import android.support.test.espresso.ViewInteraction
 import android.support.test.espresso.action.ViewActions
