+++ app/src/androidTest/java/org/mozilla/tv/firefox/ui/robots/SettingsRobot.kt
@@ -7,7 +7,7 @@ package org.mozilla.tv.firefox.ui.robots
 import android.support.test.espresso.Espresso.onView
 import android.support.test.espresso.matcher.ViewMatchers.withId
 import org.mozilla.tv.firefox.R
-import org.mozilla.tv.firefox.ext.click
+import org.mozilla.tv.firefox.helpers.ext.click
 
 /**
  * Implementation of Robot Pattern for the settings page.
